src/components/pagination/pagination.tsx
import { PaginationProps } from "../../types";

const Pagination = ({ currentPage, totalPages, onPageChange } : PaginationProps) =>  {
  const handlePageChange = (page: number) => {
    if (page >= 1 && page <= totalPages) {
      onPageChange(page);
    }
  };

  return (
    <div className="pagination flex justify-center items-center space-x-2 mt-4">
      <button
        onClick={() => handlePageChange(currentPage - 1)}
        disabled={currentPage === 1}
        className="px-4 py-2 bg-gray-300 hover:bg-gray-400 rounded-md"
      >
        Previous
      </button>
      <span className="text-sm text-gray-700">
        Page {currentPage} of {totalPages}
      </span>
      <button
        onClick={() => handlePageChange(currentPage + 1)}
        disabled={currentPage === totalPages}
        className="px-4 py-2 bg-gray-300 hover:bg-gray-400 rounded-md"
      >
        Next
      </button>
    </div>
  );
};

export default Pagination;
